Added auth to booking calls

+++ pages/api/gifu/delete.ts
@@ -1,10 +1,16 @@
 import { NextApiRequest, NextApiResponse } from 'next'
 import prisma from 'lib/prisma'
 import { GetInFitUpDTO } from 'interfaces'
+import { getEmailFromReq, checkAccess } from 'services/userService'
 
 export default async function handle (req: NextApiRequest, res: NextApiResponse) {
   try {
     const gifu = req.body as GetInFitUpDTO
+
+    const email = await getEmailFromReq(req)
+    const access = await checkAccess(email, { GifuId: gifu.Id })
+    if (!access) return res.status(401)
+
     await prisma.$transaction([
       prisma.getInFitUp.delete({
         where: {
