+++ pages/api/performances/delete.ts
@@ -1,10 +1,16 @@
 import { NextApiRequest, NextApiResponse } from 'next'
 import prisma from 'lib/prisma'
 import { PerformanceDTO } from 'interfaces'
+import { getEmailFromReq, checkAccess } from 'services/userService'
 
 export default async function handle (req: NextApiRequest, res: NextApiResponse) {
   try {
     const perf = req.body as PerformanceDTO
+
+    const email = await getEmailFromReq(req)
+    const access = await checkAccess(email, { PerformanceId: perf.Id })
+    if (!access) return res.status(401)
+
     await prisma.$transaction([
       prisma.performance.delete({
         where: {
